Test files were created for prototyping a method to make use of the shell script provided by EarthData to download the data files.
I attempted to expand the subprocess call such that the function could read inputs from a txt file with urls and download their respective files into the correct directory.

This did not work as expected.

Thankfully EarthData has an example to download the datafiles with python. This helps me avoid fumbling with shell commands.